Set window size to 1024x576. Added PlayingScreen.java which contains the background map the game will be played on.

+++ Coding Project/core/src/com/mygdx/game/TruckDriver.java
@@ -5,7 +5,7 @@ import java.awt.KeyboardFocusManager;
 import java.awt.event.KeyEvent;
 
 public class TruckDriver {
-    double fuel = 20;
+    int fuel = 20;
     int added_fuel = 0;
     int truckX = 50;
     int truckY = 100;
